+++ NINA.Photon.Plugin.ASA/View/ErrorStatisticsView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace NINA.Photon.Plugin.ASA.View
+{
+    /// <summary>
+    /// Interaction logic for ErrorStatisticsView.xaml
+    /// </summary>
+    public partial class ErrorStatisticsView : UserControl
+    {
+        public ErrorStatisticsView()
+        {
+            InitializeComponent();
+        }
+    }
+}